Refactor CRD model package structure
