+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/cell/DoneableCell.java
@@ -16,15 +16,18 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.crds;
+package io.cellery.observability.k8s.client.crds.cell;
 
-import io.fabric8.kubernetes.client.CustomResourceList;
+import io.fabric8.kubernetes.api.builder.Function;
+import io.fabric8.kubernetes.client.CustomResourceDoneable;
 
 /**
- * This class represents the List resource for Cellery cells.
+ * This class represents Doneable resource for Cellery cells.
  */
-public class CellList extends CustomResourceList<Cell> {
+public class DoneableCell extends CustomResourceDoneable<CellImpl> {
 
-    private static final long serialVersionUID = 1L;
+    public DoneableCell(CellImpl resource, Function<CellImpl, CellImpl> function) {
+        super(resource, function);
+    }
 
 }
