+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/composite/DoneableComposite.java
@@ -16,17 +16,17 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.crds;
+package io.cellery.observability.k8s.client.crds.composite;
 
 import io.fabric8.kubernetes.api.builder.Function;
 import io.fabric8.kubernetes.client.CustomResourceDoneable;
 
 /**
- * This class represents Doneable resource for Cellery cells.
+ * This class represents Doneable resource for Cellery composites.
  */
-public class DoneableCell extends CustomResourceDoneable<CellImpl> {
+public class DoneableComposite extends CustomResourceDoneable<CompositeImpl> {
 
-    public DoneableCell(CellImpl resource, Function<CellImpl, CellImpl> function) {
+    public DoneableComposite(CompositeImpl resource, Function<CompositeImpl, CompositeImpl> function) {
         super(resource, function);
     }
