+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/service/ServicesTemplate.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.crds.model;
+package io.cellery.observability.k8s.client.crds.service;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
