+++ components/global/core/io.cellery.observability.k8s.client/src/test/java/io/cellery/observability/k8s/client/UtilsTestCase.java
@@ -18,11 +18,11 @@
 
 package io.cellery.observability.k8s.client;
 
-import io.cellery.observability.k8s.client.crds.Cell;
-import io.cellery.observability.k8s.client.crds.Composite;
-import io.cellery.observability.k8s.client.crds.model.GRPC;
-import io.cellery.observability.k8s.client.crds.model.HTTP;
-import io.cellery.observability.k8s.client.crds.model.TCP;
+import io.cellery.observability.k8s.client.crds.cell.Cell;
+import io.cellery.observability.k8s.client.crds.composite.Composite;
+import io.cellery.observability.k8s.client.crds.gateway.GRPC;
+import io.cellery.observability.k8s.client.crds.gateway.HTTP;
+import io.cellery.observability.k8s.client.crds.gateway.TCP;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.api.model.Pod;
 import org.mockito.Mockito;
